Added a JS version of the client app
With the new JS version of the client app, the TS version has to be
renamed. Additionally, the package.json dependencies were updated
in the existing TS client code.

create-react-app, which was used to initialize the client apps has
begun using node-gyp & Python during the installation. Due to this
the Dockerfiles for both client apps had to be updated to install
Python into the node alpine images.